+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.9.7
+### Added
+- More debugging.
+
+### Change
+- Be defensive against duplicate peers.
+- Increase interval between housekeeping pings.
+- Replace peers in database atomically.
+
 ## 0.9.6
 ### Change
 - Randomize known peers instead of keeping scores.